Ændret placering af Nav componenter
