+++ src/components/Nav/footer/NavSocials.tsx
@@ -1,0 +1,46 @@
+export default function NavSocials() {
+    return (
+        <ul className="navbar-nav d-flex">
+            <li className="nav-item">
+                <a
+                    className={"nav-link text-secondary-emphasis"}
+                    href="https://www.facebook.com"
+                    target="_blank"
+                    rel="noreferrer"
+                >
+                    Facebook
+                </a>
+            </li>
+            <li className="nav-item">
+                <a
+                    className={"nav-link text-secondary-emphasis"}
+                    href="https://www.twitter.com"
+                    target="_blank"
+                    rel="noreferrer"
+                >
+                    Twitter
+                </a>
+            </li>
+            <li className="nav-item">
+                <a
+                    className={"nav-link text-secondary-emphasis"}
+                    href="https://www.instagram.com"
+                    target="_blank"
+                    rel="noreferrer"
+                >
+                    Instagram
+                </a>
+            </li>
+            <li className="nav-item">
+                <a
+                    className={"nav-link text-secondary-emphasis"}
+                    href="https://www.facebook.com"
+                    target="_blank"
+                    rel="noreferrer"
+                >
+                    Facebook
+                </a>
+            </li>
+        </ul>
+    );
+}
